move types into own package
