+++ packages/cosmiconfig/tsconfig.json
@@ -7,6 +7,7 @@
   "include": ["src", "test"],
   "references": [
     { "path": "../base/tsconfig.build.json" },
+    { "path": "../types/tsconfig.build.json" },
     { "path": "../yaml/tsconfig.build.json" }
   ]
 }
