+++ packages/types/tsconfig.build.json
@@ -5,6 +5,5 @@
     "rootDir": "./src",
     "outDir": "./dist"
   },
-  "include": ["src"],
-  "references": [{ "path": "../types/tsconfig.build.json" }]
+  "include": ["src"]
 }
