+++ packages/types/tsconfig.json
@@ -4,6 +4,5 @@
     "composite": false,
     "rootDir": "."
   },
-  "include": ["src", "test"],
-  "references": [{ "path": "../types/tsconfig.build.json" }]
+  "include": ["src"]
 }
